feat: two step build to export script as string
