+++ packages/theme-script/package.json
@@ -2,14 +2,19 @@
   "name": "@hiogawa/theme-script",
   "version": "0.0.0",
   "type": "module",
-  "main": "./dist/index.global.js",
-  "module": "./dist/index.global.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/index.global.js",
-      "require": "./dist/index.global.js",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs",
       "types": "./dist/index.d.ts"
+    },
+    "./dist/main": {
+      "import": "./dist/main.js",
+      "require": "./dist/main.cjs",
+      "types": "./dist/main.d.ts"
     }
   },
   "files": [
@@ -22,8 +27,11 @@
     "directory": "packages/theme-script"
   },
   "scripts": {
-    "build": "tsup",
+    "build": "BUILD_STEP=0 tsup && BUILD_STEP=1 tsup",
     "test": "vitest",
     "release": "pnpm publish --no-git-checks --access public"
+  },
+  "devDependencies": {
+    "@types/node": "^18"
   }
 }
